LayoutTests/platform/ios-sim-deprecated/accessibility/link-with-images-text-expected.txt

  iPhoneFrom $99
This test makes sure that a link with images and text appears as a link.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS obj.iphoneIsElement is true
FAIL obj.iphoneLabel should be AXLabel:  iPhoneFrom $99. Was AXLabel: iPhone From $99.
PASS successfullyParsed is true

TEST COMPLETE
